requirements.txt
# Django framework
Django==4.2.7

# Image processing
Pillow==10.0.1

# Forms and admin enhancements
django-crispy-forms==2.0
django-summernote==0.8.20.0
django-taggit==4.0.0
django-seo==0.3.5

# Static file serving
whitenoise==6.6.0

# WSGI server
gunicorn==21.2.0

# Database driver
psycopg2-binary==2.9.7

# Environment variable management
python-decouple==3.8
python-dotenv>=1.0.0

# Utilities
setuptools_scm>=7.0
requests>=2.31.0
